manifestation only works when youre above a 7
