+++ packages/synapse-interface/components/Pools/PoolActionOptions.tsx
@@ -8,7 +8,7 @@ import { UnstakeIcon } from '../icons/UnstakeIcon'
 import { WithdrawIcon } from '../icons/WithdrawIcon'
 import { ClaimIcon } from '../icons/ClaimIcon'
 import { DownArrow } from '../icons/DownArrow'
-
+import { OptionButton } from "@/components/buttons/OptionButton"
 export const PoolActionOptions = ({
   pool,
   options,
@@ -145,23 +145,4 @@ export function TransactionPopoverContainer({
   )
 }
 
-export const OptionButton = ({
-  icon,
-  text,
-  onClick,
-}: {
-  icon: any
-  text: string
-  onClick: () => void
-}) => {
-  return (
-    <div
-      data-test-id="option-button"
-      onClick={onClick}
-      className="flex hover:cursor-pointer hover:bg-slate-400/20 rounded-sm p-1 text-white/80 hover:text-white"
-    >
-      <div className="my-auto mr-1">{icon}</div>
-      <div className="text-sm">{text}</div>
-    </div>
-  )
-}
+
