+++ packages/synapse-interface/components/StateManagedBridge/DestinationAddressInput.tsx
@@ -13,23 +13,22 @@ export const DestinationAddressInput = ({
   const dispatch = useDispatch()
   const chain = CHAINS_BY_ID[toChainId]
   const chainName = chain?.name || 'nulls'
-  let placeholder
-  placeholder = `Enter ${chainName} address...`
+  const placeholder = `Enter ${chainName} address...`
 
   return (
     <div>
       <div className="w-[30%]">
-        <div className="flex items-center justify-center  h-[26px] -mt-4 p-2 absolute ml-5 md:ml-10 text-sm text-[#D8D1DC] rounded-md bg-bgLight">
+        <div className="flex items-center justify-center  h-[26px] -mt-4 p-2 absolute ml-5 md:ml-10 text-sm text-[#D8D1DC] rounded-md bg-slate-700">
           Withdraw to...
         </div>
       </div>
-      <div className="h-16 px-2 pb-4 mt-4 space-x-2 text-left sm:px-5">
+      <div className="h-16 pb-4 mt-4 text-left sm:px-2">
         <div
           className={`
             h-14 flex flex-grow items-center
-            bg-transparent
-            border border-bgLight hover:border-bgLightest focus-within:border-bgLightest
-            pl-3 sm:pl-4
+            bg-bgBase/10
+            border border-white/10 hover:border-white/50 focus-within:border-white/50
+            pl-3 sm:pl-3
             py-0.5 rounded-md
           `}
         >
@@ -42,9 +41,9 @@ export const DestinationAddressInput = ({
               bg-transparent
               w-[300px]
               sm:min-w-[400px]
-              max-w-[calc(100%-92px)]
+              max-w-[calc(100%-88px)]
               sm:w-full
-              text-white text-opacity-80 text-lg
+              text-white text-opacity-80
               placeholder:text-[#88818C]
             `}
             placeholder={placeholder}
