+++ packages/synapse-interface/components/StateManagedBridge/FromTokenListOverlay.tsx
@@ -15,8 +15,8 @@ import { useBridgeState } from '@/slices/bridge/hooks'
 import SelectSpecificTokenButton from './components/SelectSpecificTokenButton'
 import { getRoutePossibilities } from '@/utils/routeMaker/generateRoutePossibilities'
 
-import { hasBalance } from './helpers/sortByBalance'
-import { sortByPriorityRank } from './helpers/sortByPriorityRank'
+import { hasBalance } from '@/utils/helpers/hasBalance'
+import { sortByPriorityRank } from '@/utils/helpers/sortByPriorityRank'
 import { CHAINS_BY_ID } from '@/constants/chains'
 import useCloseOnOutsideClick from '@/utils/hooks/useCloseOnOutsideClick'
 import { CloseButton } from './components/CloseButton'
