+++ packages/synapse-interface/components/StateManagedSwap/SwapFromTokenListOverlay.tsx
@@ -11,8 +11,8 @@ import { segmentAnalyticsEvent } from '@/contexts/SegmentAnalyticsProvider'
 import { usePortfolioBalances } from '@/slices/portfolio/hooks'
 import SelectSpecificTokenButton from './components/SelectSpecificTokenButton'
 
-import { hasBalance } from './helpers/sortByBalance'
-import { sortByPriorityRank } from './helpers/sortByPriorityRank'
+import { hasBalance } from '@/utils/helpers/hasBalance'
+import { sortByPriorityRank } from '@/utils/helpers/sortByPriorityRank'
 import useCloseOnOutsideClick from '@/utils/hooks/useCloseOnOutsideClick'
 import { CloseButton } from './components/CloseButton'
 import { SearchResults } from './components/SearchResults'
