+++ packages/synapse-interface/components/_Transaction/components/MenuItem.tsx
@@ -1,3 +1,4 @@
+import { OptionButton } from "@/components/buttons/OptionButton"
 export const MenuItem = ({
   text,
   link,
@@ -38,26 +39,5 @@ export const MenuItem = ({
         </a>
       )}
     </li>
-  )
-}
-
-export const OptionButton = ({
-  icon,
-  text,
-  onClick,
-}: {
-  icon?: any
-  text: string
-  onClick: () => void
-}) => {
-  return (
-    <div
-      data-test-id="option-button"
-      onClick={onClick}
-      className="flex hover:cursor-pointer hover:bg-slate-400/20 rounded-sm p-1 text-white/80 hover:text-white"
-    >
-      <div className="my-auto mr-1">{icon}</div>
-      <div className="text-sm">{text}</div>
-    </div>
   )
 }
