+++ packages/synapse-interface/utils/helpers/hasBalance.ts
@@ -12,10 +12,12 @@ export const hasBalance = (
     return false
   }
   const pb = portfolioBalances[chainId]
-  const token = _(pb)
-    .pickBy((value, _key) => value.token === t)
-    .value()
+  if (!pb) {
+    return false
+  }
+  const token = _.pickBy(pb, (value, _key) => value.token === t)
+
   const tokenWithPb = Object.values(token)[0]
 
-  return tokenWithPb?.balance !== 0n
+  return tokenWithPb && tokenWithPb.balance !== 0n
 }
