--- packages/synapse-interface/utils/remove0xPrefix.ts
@@ -1,9 +1,0 @@
-export const remove0xPrefix = (str?: string): string => {
-  if (!str) {
-    return str
-  }
-  if (str.startsWith('0x')) {
-    return str.slice(2)
-  }
-  return str
-}
